refactor: rename comp to PaperTitle
